refactor: Models와 Services 분리
