+++ src/InterviewAssistant.ApiService/Models/ChatRequest.cs
@@ -1,0 +1,10 @@
+using System.Text.Json.Serialization;
+
+namespace InterviewAssistant.ApiService.Models;
+
+
+public class ChatRequest
+{
+    [JsonPropertyName("message")]
+    public string Message { get; set; } = string.Empty;
+}
