+++ src/InterviewAssistant.ApiService/Models/ChatResponse.cs
@@ -1,0 +1,12 @@
+using System.Text.Json.Serialization;
+
+namespace InterviewAssistant.ApiService.Models;
+
+public class ChatResponse
+{
+    [JsonPropertyName("response")]
+    public string Response { get; set; } = string.Empty;
+    
+    [JsonPropertyName("timestamp")]
+    public DateTime Timestamp { get; set; } = DateTime.UtcNow;
+}
