+++ src/InterviewAssistant.ApiService/Program.cs
@@ -1,15 +1,15 @@
-using Microsoft.AspNetCore.OpenApi; // 추가된 네임스페이스
+using Microsoft.AspNetCore.OpenApi;
 using Microsoft.SemanticKernel;
 using Microsoft.SemanticKernel.ChatCompletion;
 using Microsoft.SemanticKernel.Connectors.OpenAI;
-using System.Text.Json.Serialization;
-using System.Threading;
+using InterviewAssistant.ApiService.Models;
+using InterviewAssistant.ApiService.Services;
 
 var builder = WebApplication.CreateBuilder(args);
 
 // API 탐색기 사용 (.NET 9의 기본 OpenAPI 패키지)
 builder.Services.AddEndpointsApiExplorer();
-builder.Services.AddSwaggerGen(); // .NET 9에서도 Swagger 패키지는 여전히 필요합니다
+builder.Services.AddSwaggerGen();
 
 // Semantic Kernel 서비스 등록
 builder.Services.AddSingleton(sp =>
@@ -31,18 +31,6 @@ builder.Services.AddSingleton(sp =>
     return kernelBuilder.Build();
 });
 
-// CORS 설정 추가
-builder.Services.AddCors(options =>
-{
-    options.AddPolicy("AllowAll", builder =>
-    {
-        builder
-            .AllowAnyOrigin()
-            .AllowAnyMethod()
-            .AllowAnyHeader();
-    });
-});
-
 // ServiceDefaults 추가 (.NET Aspire 프로젝트에 포함된 기본 설정)
 builder.AddServiceDefaults();
 
@@ -61,9 +49,6 @@ if (app.Environment.IsDevelopment())
     });
 }
 
-// CORS 미들웨어 활성화
-app.UseCors("AllowAll");
-
 // 기타 미들웨어 설정
 app.UseHttpsRedirection();
 
@@ -75,6 +60,7 @@ chatGroup.MapPost("/", async (ChatRequest request, Kernel kernel, ChatState stat
 {
     try
     {
+        // 빈 메시지 확인
         if (string.IsNullOrEmpty(request.Message))
         {
             return Results.BadRequest("메시지 내용이 비어있습니다.");
@@ -84,7 +70,7 @@ chatGroup.MapPost("/", async (ChatRequest request, Kernel kernel, ChatState stat
         await state.ApplyRateLimiting(logger);
 
         // 채팅 히스토리 생성
-        var chatHistory = new Microsoft.SemanticKernel.ChatCompletion.ChatHistory();
+        var chatHistory = new ChatHistory();
         chatHistory.AddSystemMessage("당신은 유용한 AI 챗봇입니다. 짧고 명확하게 답변해주세요.");
         
         // 이전 대화가 있으면 추가
@@ -101,9 +87,9 @@ chatGroup.MapPost("/", async (ChatRequest request, Kernel kernel, ChatState stat
         // 짧은 응답을 위한 설정
         var openAIPromptExecutionSettings = new OpenAIPromptExecutionSettings
         {
-            MaxTokens = 200,
-            Temperature = 0.7,
-            TopP = 0.95
+            MaxTokens = 200,  // 최대 토큰 수 제한
+            Temperature = 0.7, // 창의성 조절 (0: 결정적, 1: 창의적)
+            TopP = 0.95 // 다양성 조절
         };
 
         string responseText;
@@ -121,11 +107,12 @@ chatGroup.MapPost("/", async (ChatRequest request, Kernel kernel, ChatState stat
         }
         catch (Exception ex) 
         {
+            // 오류 발생 시 로깅 및 기본 응답 제공
             logger.LogError(ex, "Error calling AI service");
-            // 실패 시 간단한 응답 제공
             responseText = "죄송합니다. 현재 시스템이 바쁩니다. 잠시 후 다시 시도해주세요.";
         }
         
+        // 마지막 대화 저장
         state.LastBotResponse = responseText;
 
         // 응답 생성
@@ -140,6 +127,7 @@ chatGroup.MapPost("/", async (ChatRequest request, Kernel kernel, ChatState stat
     }
     catch (Exception ex)
     {
+        // 예상치 못한 오류 처리
         logger.LogError(ex, "Chat request processing failed");
         return Results.Problem($"내부 서버 오류: {ex.Message}");
     }
@@ -160,58 +148,3 @@ chatGroup.MapPost("/reset", (ChatState state) =>
 app.MapDefaultEndpoints();
 
 app.Run();
-
-// 데이터 모델 정의
-public class ChatRequest
-{
-    [JsonPropertyName("message")]
-    public string Message { get; set; } = string.Empty;
-}
-
-public class ChatResponse
-{
-    [JsonPropertyName("response")]
-    public string Response { get; set; } = string.Empty;
-    
-    [JsonPropertyName("timestamp")]
-    public DateTime Timestamp { get; set; } = DateTime.UtcNow;
-}
-
-// 채팅 상태 관리를 위한 싱글톤 서비스
-public class ChatState
-{
-    // 간단한 히스토리 (최근 메시지 2개만 유지)
-    public string LastUserMessage { get; set; } = string.Empty;
-    public string LastBotResponse { get; set; } = string.Empty;
-    
-    // 속도 제한을 위한 타임스탬프
-    public DateTime LastRequestTime { get; private set; } = DateTime.MinValue;
-    private readonly SemaphoreSlim _throttleLock = new SemaphoreSlim(1, 1);
-    
-    public async Task ApplyRateLimiting(ILogger logger)
-    {
-        await _throttleLock.WaitAsync();
-        try
-        {
-            var timeSinceLastRequest = DateTime.UtcNow - LastRequestTime;
-            if (timeSinceLastRequest.TotalSeconds < 60)
-            {
-                var waitTime = TimeSpan.FromSeconds(60) - timeSinceLastRequest;
-                logger.LogInformation("Rate limiting: Waiting for {WaitTime}ms before next request", waitTime.TotalMilliseconds);
-                await Task.Delay(waitTime);
-            }
-            
-            LastRequestTime = DateTime.UtcNow;
-        }
-        finally
-        {
-            _throttleLock.Release();
-        }
-    }
-    
-    public void Reset()
-    {
-        LastUserMessage = string.Empty;
-        LastBotResponse = string.Empty;
-    }
-}
